+++ packages/toolpad-app/src/appDom.ts
@@ -1,16 +1,16 @@
 import { generateKeyBetween } from 'fractional-indexing';
 import cuid from 'cuid';
-import { ArgTypeDefinitions, PropValueType, PropValueTypes } from '@mui/toolpad-core';
 import {
-  NodeId,
   ConstantAttrValue,
   BindableAttrValue,
   BindableAttrValues,
-  ConnectionStatus,
-  AppTheme,
   ConstantAttrValues,
   SecretAttrValue,
-} from './types';
+  ArgTypeDefinitions,
+  PropValueType,
+  PropValueTypes,
+} from '@mui/toolpad-core';
+import { NodeId, ConnectionStatus, AppTheme } from './types';
 import { omit, update, updateOrCreate } from './utils/immutability';
 import { camelCase, generateUniqueString, removeDiacritics } from './utils/strings';
 import { ExactEntriesOf } from './utils/types';
