+++ packages/toolpad-app/src/components/AppEditor/PageEditor/BindableEditor.tsx
@@ -1,9 +1,14 @@
 import { Alert, Stack } from '@mui/material';
 import * as React from 'react';
-import { ArgTypeDefinition, ArgControlSpec, PropValueType } from '@mui/toolpad-core';
+import {
+  BindableAttrValue,
+  ArgTypeDefinition,
+  ArgControlSpec,
+  PropValueType,
+} from '@mui/toolpad-core';
 import propertyControls from '../../propertyControls';
 import { BindingEditor } from '../BindingEditor';
-import { NodeId, BindableAttrValue } from '../../../types';
+import { NodeId } from '../../../types';
 import { WithControlledProp } from '../../../utils/types';
 import { usePageEditorState } from './PageEditorProvider';
